fix(worker): prevent checkin of non-existing pids
Previously, if a connection was terminated and the client tried to
check it back in, the worker would create a new record for the dead
connection and add it to the pool. This polluted the pool with dead
connections, leading to errors.

This commit changes conn_checkin/3 to return an error if the pid is
not found in the ets table, preventing dead connections from being
re-introduced.